requirements.txt
paddlepaddle==2.6.0
paddleocr==2.7.3
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3